feat: Enhance student management and import functionality
- Added a new endpoint to retrieve inactive students by department in studentControllers.js.
- Implemented Excel file import for students, creating accounts and handling errors in studentControllers.js.
- Updated Swagger documentation for student registration and import endpoints.
- Added status filtering for approved scientific papers in statisticsController.js.
- Improved user work update logic in userWorkController.js.
- Refactored work unit update logic in workUnitController.js.
- Modified Messages model to remove unnecessary comments and ensure clarity.
- Updated ScientificPaper model to make certain fields optional.
- Added timestamps to ScoringFormula model.
- Enhanced Student model to include a new degree option.
- Adjusted UserWork model to allow for multiple user references.
- Updated authRoutes.js to include new registration and approval endpoints with Swagger documentation.
- Added import functionality for lecturers from Excel files in lecturersRoutes.js.
- Removed paperViews.js routes and consolidated functionality into paperViewsRoutes.js.
- Added new endpoints to retrieve paper downloads and views by user in respective routes.
- Updated searchAIRoutes.js to remove unnecessary department filter.

+++ build/src/controller/attributeController.js
@@ -1,6 +1,6 @@
 const Attribute = require("../models/Attribute");
+const ScoringFormula = require("../models/ScoringFormula");
 const attributeController = {
-  // Tạo mới một Attribute
   createAttribute: async (req, res) => {
     try {
       const {
@@ -11,13 +11,20 @@ const attributeController = {
       });
       if (existingAttribute) {
         return res.status(400).json({
-          message: "Attribute name already exists"
+          message: "Thuộc tính đã tồn tại"
         });
       }
       const attribute = new Attribute(req.body);
       await attribute.save();
       res.status(201).json(attribute);
     } catch (error) {
+      if (error.name === "ValidationError") {
+        const errors = Object.values(error.errors).map(err => err.message);
+        return res.status(400).json({
+          message: "Lỗi xác thực",
+          errors
+        });
+      }
       res.status(400).json({
         message: error.message
       });
@@ -89,7 +96,17 @@ const attributeController = {
     try {
       const {
         name
-      } = req.body;
+      } = req.params;
+
+      // Kiểm tra xem thuộc tính có đang được dùng trong công thức điểm không
+      const formulaUsingAttribute = await ScoringFormula.findOne({
+        attributeName: name
+      });
+      if (formulaUsingAttribute) {
+        return res.status(400).json({
+          message: "Không thể xóa thuộc tính vì đang được sử dụng trong công thức điểm."
+        });
+      }
       const attribute = await Attribute.findOneAndDelete({
         name
       });
